moved functions into node-core from utils
